+++ crates/thumbsup-shell/src/logging.rs
@@ -106,7 +106,7 @@ fn epoch_to_civil(secs: u64) -> (i32, u32, u32, u32, u32, u32) {
     let era = z.div_euclid(146097);
     let doe = z.rem_euclid(146097) as u64;
     let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146096) / 365;
-    let y = (yoe as i64 + era * 400) as i64;
+    let y = yoe as i64 + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
     let mp = (5 * doy + 2) / 153;
     let d = (doy - (153 * mp + 2) / 5 + 1) as u32;
